@@ -30,11 +30,11 @@ struct ColorCellViewModelGrid: CellViewModel {
 
     // MARK: Hashable
 
-    func hash(into hasher: inout Hasher) {
+    nonisolated func hash(into hasher: inout Hasher) {
         hasher.combine(self.color)
     }
 
-    static func == (left: Self, right: Self) -> Bool {
+    nonisolated static func == (left: Self, right: Self) -> Bool {
         left.color == right.color
     }
 }
