@@ -17,89 +17,50 @@ import UIKit
 
 final class ListViewController: ExampleViewController, CellEventCoordinator {
 
-    var cancellables = [AnyCancellable]()
+    lazy var driver = CollectionViewDriver(
+        view: self.collectionView,
+        layout: self.makeLayout(),
+        emptyViewProvider: sharedEmptyViewProvider,
+        cellEventCoordinator: self
+    ) { [unowned self] driver in
+        print("list did update!")
+        print(driver.viewModel)
+    }
 
     override var model: Model {
         didSet {
             // Every time the model updates, regenerate and set the view model
-            self.driver.viewModel = self.makeCollectionViewModel()
+            self.driver.viewModel = self.makeViewModel()
         }
     }
+
+    private var cancellables = [AnyCancellable]()
 
     // MARK: CellEventCoordinator
 
-    func didSelectCell(viewModel: any CellViewModel) {
-        print("\(#function): \(viewModel.id)")
-
-        if let personVM = viewModel as? PersonCellViewModelList {
-            let personVC = PersonViewController(person: personVM.person)
-            self.navigationController?.pushViewController(personVC, animated: true)
-            return
-        }
-
-        if let colorVM = viewModel as? ColorCellViewModelList {
-            let colorVC = ColorViewController(color: colorVM.color)
-            self.navigationController?.pushViewController(colorVC, animated: true)
-            return
-        }
-
-        assertionFailure("unhandled cell selection")
-    }
+    // In this example, the cell view models handle cell selection and navigation themselves.
 
     // MARK: View lifecycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        let layout = self.makeLayout()
-        let viewModel = self.makeCollectionViewModel()
-
-        self.driver = CollectionViewDriver(
-            view: self.collectionView,
-            layout: layout,
-            viewModel: viewModel,
-            cellEventCoordinator: self,
-            animateUpdates: true
-        ) { [unowned self] in
-            print("list did update!")
-            print(self.driver.viewModel)
-        }
+        self.driver.viewModel = self.makeViewModel()
 
         self.driver.$viewModel
             .sink { _ in
                 print("did publish view model update")
             }
             .store(in: &self.cancellables)
-
     }
 
-    func makeLayout() -> UICollectionViewCompositionalLayout {
+    // MARK: Private
+
+    private func makeLayout() -> UICollectionViewCompositionalLayout {
         UICollectionViewCompositionalLayout { _, layoutEnvironment in
             var configuration = UICollectionLayoutListConfiguration(appearance: .insetGrouped)
             configuration.headerMode = .supplementary
             configuration.footerMode = .supplementary
-
-            // TODO: swipe actions need re-working. need to reference item identifier, not index path.
-//            configuration.leadingSwipeActionsConfigurationProvider = { [unowned self] indexPath in
-//                let favoriteAction = UIContextualAction(style: .normal, title: "Favorite") { _, _, completion in
-//                    // self.toggleFavoriteAt(indexPath: indexPath)
-//                    completion(true)
-//                }
-//                favoriteAction.image = UIImage(systemName: "star.fill")
-//                favoriteAction.backgroundColor = .systemYellow
-//                return UISwipeActionsConfiguration(actions: [favoriteAction])
-//            }
-//
-//            configuration.trailingSwipeActionsConfigurationProvider = { [unowned self] indexPath in
-//                let deleteAction = UIContextualAction(style: .destructive, title: "Delete") { _, _, completion in
-//                    // self.deleteAt(indexPath: indexPath)
-//                    completion(true)
-//                }
-//                deleteAction.image = UIImage(systemName: "trash")
-//                deleteAction.backgroundColor = .systemRed
-//                return UISwipeActionsConfiguration(actions: [deleteAction])
-//            }
-
             return NSCollectionLayoutSection.list(
                 using: configuration,
                 layoutEnvironment: layoutEnvironment
@@ -107,7 +68,7 @@ final class ListViewController: ExampleViewController, CellEventCoordinator {
         }
     }
 
-    func makeCollectionViewModel() -> CollectionViewModel {
+    private func makeViewModel() -> CollectionViewModel {
         // Create People Section
         let peopleCellViewModels = self.model.people.map {
             let menuConfig = UIContextMenuConfiguration.configFor(
@@ -123,7 +84,7 @@ final class ListViewController: ExampleViewController, CellEventCoordinator {
             return PersonCellViewModelList(
                 person: $0,
                 contextMenuConfiguration: menuConfig
-            ).anyViewModel
+            ).eraseToAnyViewModel()
         }
         let peopleHeader = HeaderViewModel(title: "People", style: .small)
         let peopleFooter = FooterViewModel(text: "\(self.model.people.count) people")
@@ -148,7 +109,7 @@ final class ListViewController: ExampleViewController, CellEventCoordinator {
             return ColorCellViewModelList(
                 color: $0,
                 contextMenuConfiguration: menuConfig
-            ).anyViewModel
+            ).eraseToAnyViewModel()
         }
         let colorHeader = HeaderViewModel(title: "Colors", style: .small)
         let colorFooter = FooterViewModel(text: "\(self.model.colors.count) colors")
@@ -160,6 +121,6 @@ final class ListViewController: ExampleViewController, CellEventCoordinator {
         )
 
         // Create final view model
-        return CollectionViewModel(sections: [peopleSection, colorSection])
+        return CollectionViewModel(id: "list_view", sections: [peopleSection, colorSection])
     }
 }
