@@ -17,9 +17,33 @@ import UIKit
 final class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
-    func application(_ application: UIApplication,
-                     // swiftlint:disable:next discouraged_optional_collection
-                     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        true
+    func application(
+        _ application: UIApplication,
+        // swiftlint:disable:next discouraged_optional_collection
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
+    ) -> Bool {
+        let grid = GridViewController()
+        grid.title = "Grid"
+        grid.tabBarItem.image = UIImage(systemName: "square.grid.2x2.fill")
+
+        let list = ListViewController()
+        list.title = "List"
+        list.tabBarItem.image = UIImage(systemName: "list.dash")
+
+        let simple = SimpleStaticViewController()
+        simple.title = "Simple Static"
+        simple.tabBarItem.image = UIImage(systemName: "list.bullet.rectangle.portrait.fill")
+
+        let tabBar = UITabBarController()
+        tabBar.viewControllers = [
+            UINavigationController(rootViewController: grid),
+            UINavigationController(rootViewController: list),
+            UINavigationController(rootViewController: simple)
+        ]
+
+        self.window = UIWindow()
+        self.window?.rootViewController = tabBar
+        self.window?.makeKeyAndVisible()
+        return true
     }
 }
