@@ -15,24 +15,39 @@ import Foundation
 import UIKit
 
 /// Defines a view model that describes and configures a cell in the collection view.
+@MainActor
 public protocol CellViewModel: DiffableViewModel, ViewRegistrationProvider {
     /// The type of cell that this view model represents and configures.
     associatedtype CellType: UICollectionViewCell
 
+    /// Returns whether or not the cell should get highlighted.
+    /// This corresponds to the delegate method `collectionView(_:shouldHighlightItemAt:)`.
+    /// The default implementation returns `true`.
     var shouldHighlight: Bool { get }
 
+    /// Returns a context menu configuration for the cell.
+    /// This corresponds to the delegate method `collectionView(_:contextMenuConfigurationForItemAt:point:)`.
     var contextMenuConfiguration: UIContextMenuConfiguration? { get }
 
+    /// Configures the provided cell for display in the collection.
+    /// - Parameter cell: The cell to configure.
     func configure(cell: CellType)
 
+    /// Handles the selection event for this cell, optionally using the provided `coordinator`.
+    /// - Parameter coordinator: An event coordinator object, if one was provided to the `CollectionViewDriver`.
     func didSelect(with coordinator: CellEventCoordinator?)
 }
 
 extension CellViewModel {
+    /// Default implementation. Returns `true`.
     public var shouldHighlight: Bool { true }
 
+    /// Default implementation. Returns `nil`.
     public var contextMenuConfiguration: UIContextMenuConfiguration? { nil }
 
+    /// Default implementation.
+    /// Calls `didSelectCell(viewModel:)` on the `coordinator`,
+    /// passing `self` to the `viewModel` parameter.
     public func didSelect(with coordinator: CellEventCoordinator?) {
         coordinator?.didSelectCell(viewModel: self)
     }
@@ -43,7 +58,7 @@ extension CellViewModel {
     public var cellClass: AnyClass { CellType.self }
 
     /// A default reuse identifier for cell registration.
-    /// Value defaults to the name of the class implementing the `CellViewModel` protocol.
+    /// Returns the name of the class implementing the `CellViewModel` protocol.
     public var reuseIdentifier: String { "\(Self.self)" }
 
     /// A default registration for this view model for class-based cells.
@@ -56,11 +71,13 @@ extension CellViewModel {
     }
 
     /// Returns a type-erased version of this view model.
-    public var anyViewModel: AnyCellViewModel {
+    public func eraseToAnyViewModel() -> AnyCellViewModel {
         AnyCellViewModel(self)
     }
 
-    public func dequeueAndConfigureCellFor(collectionView: UICollectionView, at indexPath: IndexPath) -> CellType {
+    // MARK: Internal
+
+    func dequeueAndConfigureCellFor(collectionView: UICollectionView, at indexPath: IndexPath) -> CellType {
         let cell = self.registration.dequeueViewFor(collectionView: collectionView, at: indexPath) as! CellType
         self.configure(cell: cell)
         return cell
@@ -68,27 +85,38 @@ extension CellViewModel {
 }
 
 /// A type-erased cell view model.
+///
+/// - Note: When providing cells with mixed data types to a `SectionViewModel`,
+/// it is necessary to convert them to `AnyCellViewModel`.
+@MainActor
 public struct AnyCellViewModel: CellViewModel {
     // MARK: DiffableViewModel
 
-    public var id: UniqueIdentifier { self._id }
+    /// :nodoc:
+    nonisolated public var id: UniqueIdentifier { self._id }
 
     // MARK: ViewRegistrationProvider
 
+    /// :nodoc:
     public var registration: ViewRegistration { self._registration }
 
     // MARK: CellViewModel
 
+    /// :nodoc:
     public typealias CellType = UICollectionViewCell
 
+    /// :nodoc:
     public var shouldHighlight: Bool { self._shouldHighlight }
 
+    /// :nodoc:
     public var contextMenuConfiguration: UIContextMenuConfiguration? { self._contextMenuConfiguration }
 
+    /// :nodoc:
     public func configure(cell: UICollectionViewCell) {
         self._configure(cell)
     }
 
+    /// :nodoc:
     public func didSelect(with coordinator: CellEventCoordinator?) {
         self._didSelect(coordinator)
     }
@@ -105,6 +133,9 @@ public struct AnyCellViewModel: CellViewModel {
 
     // MARK: Init
 
+    /// Initializes an `AnyCellViewModel` from the provided cell view model.
+    ///
+    /// - Parameter viewModel: The view model to type-erase.
     public init<T: CellViewModel>(_ viewModel: T) {
         self._viewModel = viewModel
         self._id = viewModel.id
@@ -122,13 +153,24 @@ public struct AnyCellViewModel: CellViewModel {
 }
 
 extension AnyCellViewModel: Equatable {
-    public static func == (left: AnyCellViewModel, right: AnyCellViewModel) -> Bool {
+    /// :nodoc:
+    nonisolated public static func == (left: AnyCellViewModel, right: AnyCellViewModel) -> Bool {
         left._viewModel == right._viewModel
     }
 }
 
 extension AnyCellViewModel: Hashable {
-    public func hash(into hasher: inout Hasher) {
+    /// :nodoc:
+    nonisolated public func hash(into hasher: inout Hasher) {
         self._viewModel.hash(into: &hasher)
+    }
+}
+
+extension AnyCellViewModel: CustomDebugStringConvertible {
+    /// :nodoc:
+    nonisolated public var debugDescription: String {
+        MainActor.assumeIsolated {
+            "\(self._viewModel)"
+        }
     }
 }
