@@ -14,26 +14,71 @@
 import Foundation
 import UIKit
 
-typealias DiffableDataSource = UICollectionViewDiffableDataSource<AnyHashable, AnyHashable>
+extension AnyHashable: @unchecked Sendable { }
 
-extension DiffableDataSource {
-    typealias Snapshot = NSDiffableDataSourceSnapshot<SectionIdentifierType, ItemIdentifierType>
+struct UncheckedCompletion: @unchecked Sendable {
+    typealias Block = () -> Void
+
+    let block: Block?
+
+    init(_ block: Block?) {
+        if let block {
+            self.block = {
+                dispatchPrecondition(condition: .onQueue(.main))
+                block()
+            }
+        } else {
+            self.block = nil
+        }
+    }
+}
+
+@MainActor
+final class DiffableDataSource: UICollectionViewDiffableDataSource<AnyHashable, AnyHashable> {
+    typealias Snapshot = NSDiffableDataSourceSnapshot<AnyHashable, AnyHashable>
 
     typealias SnapshotCompletion = () -> Void
 
-    convenience init(view: UICollectionView) {
-        self.init(collectionView: view) { _, _, _ in
-            nil
-        }
-    }
+    // Avoid retaining the collection view, we know it is owned and kept alive by the driver.
+    // Thus, unowned is safe here.
+    private unowned let _collectionView: UICollectionView
 
-    convenience init(
+    private let _diffOnBackgroundQueue: Bool
+
+    private lazy var _diffingQueue = DispatchQueue(
+        label: "com.jessesquires.ReactiveCollectionsKit",
+        qos: .userInteractive,
+        autoreleaseFrequency: .workItem
+    )
+
+    // MARK: Init
+
+    init(
         view: UICollectionView,
+        diffOnBackgroundQueue: Bool,
         cellProvider: @escaping DiffableDataSource.CellProvider,
         supplementaryViewProvider: @escaping DiffableDataSource.SupplementaryViewProvider
     ) {
-        self.init(collectionView: view, cellProvider: cellProvider)
+        self._collectionView = view
+        self._diffOnBackgroundQueue = diffOnBackgroundQueue
+        super.init(collectionView: view, cellProvider: cellProvider)
         self.supplementaryViewProvider = supplementaryViewProvider
+    }
+
+    convenience init(view: UICollectionView, diffOnBackgroundQueue: Bool) {
+        self.init(
+            view: view,
+            diffOnBackgroundQueue: diffOnBackgroundQueue,
+            cellProvider: { _, _, _ in nil },
+            supplementaryViewProvider: { _, _, _ in nil }
+        )
+    }
+
+    // MARK: Applying snapshots
+
+    func reload(_ viewModel: CollectionViewModel, completion: SnapshotCompletion?) {
+        let snapshot = DiffableSnapshot(viewModel: viewModel)
+        self._applyReloadSnapshot(snapshot, completion: UncheckedCompletion(completion))
     }
 
     func applySnapshot(
@@ -45,132 +90,181 @@ extension DiffableDataSource {
         // Build initial destination snapshot, then make adjustments below.
         var destinationSnapshot = DiffableSnapshot(viewModel: destination)
 
-        // Apply item reloads, then section reloads.
-        // Reloading a section is required to properly reload headers, footers, and other supplementary views.
-        // This 2-step process is necessary to preserve collection view animations
-        // and prevent UIKit data source internal inconsistency exceptions.
+        // Apply item reconfigures, then supplementary view reconfigures.
+        // Note: "reconfigure" is a lighter weight, more efficient "reload".
         //
-        // If we only reloaded a section (instead of reloading items first),
-        // there are 2 problems:
-        //   1. item updates would not animate correctly, because the whole section is reloaded.
-        //   2. item inserts/deletes could trigger an internal inconsistency exception.
+        // Unfortunately headers, footers, and other supplementary views
+        // do not get properly reconfigured or reloaded if they display dynamic data.
+        // This is a shortcoming in UIKit. The APIs we need to do this do not exist.
+        // Ideally, `DiffableDataSource` would offer the same APIs for supplementary views
+        // as it does for items and sections.
+        //
+        // Generally speaking, hard-reloading an entire section is the only way to get
+        // headers, footers, and other supplementary views to reload/reconfigure.
+        // That sucks. Luckily, we can do better.
+        //
+        // Below is a 2-step process that is necessary to preserve collection view animations
+        // and prevent UIKit data source internal inconsistency exceptions.
 
         // Find and perform item (cell) updates first.
-        let itemsToReload = self.findItemsToReload(from: source, to: destination)
-        destinationSnapshot.reconfigureItems(itemsToReload)
+        // Add the item reconfigure updates to the snapshot.
+        let itemsToReconfigure = self._findItemsToReconfigure(from: source, to: destination)
+        destinationSnapshot.reconfigureItems(itemsToReconfigure)
 
-        // Apply snapshot with item reload updates.
-        self.applySnapshot(destinationSnapshot, animated: animated) {
-            // Once item reloads are complete, find and apply section reloads, if needed.
-            // This is necessary to update SUPPLEMENTARY VIEWS ONLY.
-            // Supplementary views do not get reloaded / reconfigured automatically when they change.
-            // To trigger updates on supplementary views, the section must be reloaded.
-            // Yes, this kinda sucks.
+        // Apply the snapshot with item reconfigure updates.
+        self._applyDiffSnapshot(destinationSnapshot, animated: animated, completion: UncheckedCompletion {
 
-            let allSourceSections = source.allSectionsByIdentifier
+            // Once the snapshot with item reconfigures is applied,
+            // we need to find and apply supplementary view reconfigures, if needed.
+            //
+            // This is necessary to update all headers, footers, and supplementary views.
+            // Per notes above, supplementary views do not get reloaded / reconfigured
+            // automatically by `DiffableDataSource` when they change.
+            //
+            // To trigger updates on supplementary views with the existing APIs,
+            // the entire section must be reloaded. Yes, that sucks. We don't want to do that.
+            // That causes all items in the section to be hard-reloaded, too.
+            // Aside from the performance impact, doing that results in an ugly UI "flash"
+            // for all item cells in the collection. Gross.
+            //
+            // However, we can actually do much better than a hard reload!
+            // Instead of reloading the entire section, we can find and compare
+            // the supplementary views and manually reconfigure them if they changed.
+            //
+            // NOTE: this only matters if supplementary views are not static.
+            // That is, if they reflect data in the data source.
+            //
+            // For example, a header with a fixed title (e.g. "My Items") will NOT need to be reloaded.
+            // However, a header that displays changing data WILL need to be reloaded.
+            // (e.g. "My 10 Items")
 
-            // Only get sections that have supplementary views.
-            let allDestinationSections = destination.allSectionsByIdentifier.filter { _, value in
-                value.hasSupplementaryViews
-            }
+            // Check all the supplementary views and reconfigure them, if needed.
+            self._reconfigureSupplementaryViewsIfNeeded(from: source, to: destination)
 
-            // If no sections have supplementary views, there's nothing to do.
-            guard allDestinationSections.isNotEmpty else {
-                completion?()
-                return
-            }
-
-            var sectionsToReload = Set<UniqueIdentifier>()
-
-            // As soon as we find 1 supplementary view in a section that needs reloading,
-            // we can mark the whole section for reload and exit early.
-            for (sectionId, destinationSection) in allDestinationSections {
-                // If this section does not exist in the source, then it is newly inserted.
-                // Thus, nothing to do.
-                guard let sourceSection = allSourceSections[sectionId] else {
-                    continue
-                }
-
-                // If this section exist in the source,
-                // and it has changed its SUPPLEMENTARY VIEWS ONLY,
-                // then only reload the section.
-                // First, check headers and footers.
-                if destinationSection.header != sourceSection.header {
-                    sectionsToReload.insert(sectionId)
-                    continue
-                }
-
-                if destinationSection.footer != sourceSection.footer {
-                    sectionsToReload.insert(sectionId)
-                    continue
-                }
-
-                // Next, check all supplementary views.
-                let allSourceSectionSupplementaryViews = sourceSection.allSupplementaryViewsByIdentifier
-
-                for destinationView in destinationSection.supplementaryViews {
-                    // If this view does not exist in the source, then it is newly added.
-                    // Thus, nothing to do.
-                    guard let sourceView = allSourceSectionSupplementaryViews[destinationView.id] else {
-                        continue
-                    }
-
-                    // After finding one view that needs reloading, we can stop,
-                    // because we have to reload the whole section anyway.
-                    if destinationView != sourceView {
-                        sectionsToReload.insert(sectionId)
-                        break
-                    }
-                }
-            }
-
-            // If no section changes, ignore and call completion
-            guard sectionsToReload.isNotEmpty else {
-                completion?()
-                return
-            }
-
-            destinationSnapshot.reloadSections(sectionsToReload.toArray)
-
-            // Apply final section updates
-            self.applySnapshot(destinationSnapshot, animated: animated, completion: completion)
-        }
+            // Finally, we're done and can call completion.
+            completion?()
+        })
     }
 
-    func findItemsToReload(
+    private func _findItemsToReconfigure(
         from source: CollectionViewModel,
         to destination: CollectionViewModel
     ) -> [UniqueIdentifier] {
-        // Determine which items need to be reloaded.
-        let allSourceCells = source.allCellsByIdentifier
-        let allDestinationCells = destination.allCellsByIdentifier
+        let allSourceCells = source.allCellsByIdentifier()
+        let allDestinationCells = destination.allCellsByIdentifier()
 
-        var itemsToReload = [UniqueIdentifier]()
+        var itemsToReconfigure = [UniqueIdentifier]()
 
         for (cellId, destinationCell) in allDestinationCells {
             let sourceCell = allSourceCells[cellId]
 
             // If this cell exist in the source, and it has changed, then reload it.
             if destinationCell != sourceCell {
-                itemsToReload.append(cellId)
+                itemsToReconfigure.append(cellId)
             }
         }
 
-        return itemsToReload
+        return itemsToReconfigure
     }
 
-    func applySnapshot(_ snapshot: Snapshot, animated: Bool, completion: SnapshotCompletion? = nil) {
-        self.apply(snapshot, animatingDifferences: animated, completion: completion)
+    private func _reconfigureSupplementaryViewsIfNeeded(
+        from source: CollectionViewModel,
+        to destination: CollectionViewModel
+    ) {
+        let allSourceSections = source.allSectionsByIdentifier()
+
+        for sectionIndex in 0..<destination.sections.count {
+            let destinationSection = destination.sections[sectionIndex]
+
+            // If this section does not have any supplementary views, skip it.
+            guard destinationSection.hasSupplementaryViews else {
+                continue
+            }
+
+            // If this section does not exist in the source,
+            // then it is newly inserted. Thus, nothing to do.
+            guard let sourceSection = allSourceSections[destinationSection.id] else {
+                continue
+            }
+
+            // This section exists in the source.
+            // Check if it has changed its supplementary views.
+
+            // First, check and reconfigure header.
+            if let destinationHeader = destinationSection.header,
+               let sourceHeader = sourceSection.header,
+               destinationHeader != sourceHeader {
+                self._reconfigureSupplementaryView(
+                    model: destinationHeader,
+                    item: 0,
+                    section: sectionIndex
+                )
+            }
+
+            // Second, check and reconfigure footer.
+            if let destinationFooter = destinationSection.footer,
+               let sourceFooter = sourceSection.footer,
+               destinationFooter != sourceFooter {
+                self._reconfigureSupplementaryView(
+                    model: destinationFooter,
+                    item: 0,
+                    section: sectionIndex
+                )
+            }
+
+            // Third, check all supplementary views.
+            let allSourceSectionSupplementaryViews = sourceSection.allSupplementaryViewsByIdentifier()
+
+            for viewIndex in 0..<destinationSection.supplementaryViews.count {
+                let destinationView = destinationSection.supplementaryViews[viewIndex]
+
+                // If this view does not exist in the source,
+                // then it is newly added. Thus, nothing to do.
+                guard let sourceView = allSourceSectionSupplementaryViews[destinationView.id] else {
+                    continue
+                }
+
+                // Check and reconfigure supplementary view.
+                if destinationView != sourceView {
+                    self._reconfigureSupplementaryView(
+                        model: destinationView,
+                        item: viewIndex,
+                        section: sectionIndex
+                    )
+                }
+            }
+        }
     }
 
-    func reload(_ viewModel: CollectionViewModel, completion: SnapshotCompletion?) {
-        let snapshot = DiffableSnapshot(viewModel: viewModel)
-        self.applySnapshotUsingReloadData(snapshot, completion: completion)
+    private func _reconfigureSupplementaryView(model: AnySupplementaryViewModel, item: Int, section: Int) {
+        let indexPath = IndexPath(item: item, section: section)
+        if let view = self._collectionView.supplementaryView(forElementKind: model._kind, at: indexPath) {
+            model.configure(view: view)
+        }
     }
-}
 
-extension Set {
-    fileprivate var toArray: [Self.Element] {
-        Array(self)
+    // MARK: Diffing
+
+    private func _applyDiffSnapshot(_ snapshot: Snapshot, animated: Bool, completion: UncheckedCompletion) {
+        self._performOnDiffingQueueIfNeeded {
+            // UIKit guarantees `completion` is called on the main queue.
+            self.apply(snapshot, animatingDifferences: animated, completion: completion.block)
+        }
+    }
+
+    func _applyReloadSnapshot(_ snapshot: Snapshot, completion: UncheckedCompletion) {
+        self._performOnDiffingQueueIfNeeded {
+            // UIKit guarantees `completion` is called on the main queue.
+            self.applySnapshotUsingReloadData(snapshot, completion: completion.block)
+        }
+    }
+
+    private func _performOnDiffingQueueIfNeeded(_ action: @Sendable @escaping () -> Void) {
+        if self._diffOnBackgroundQueue {
+            self._diffingQueue.async(execute: action)
+        } else {
+            dispatchPrecondition(condition: .onQueue(.main))
+            action()
+        }
     }
 }
