@@ -15,13 +15,23 @@ import Foundation
 import UIKit
 
 /// Describes all information needed to register a view for reuse with a `UICollectionView`.
+@MainActor
 public struct ViewRegistration: Hashable {
+    /// The view reuse identifier.
     public let reuseIdentifier: String
 
+    /// The type of view to register.
     public let viewType: ViewRegistrationViewType
 
+    /// The view registration method.
     public let method: ViewRegistrationMethod
 
+    /// Initializes a `ViewRegistration`.
+    ///
+    /// - Parameters:
+    ///   - reuseIdentifier: The view reuse identifier.
+    ///   - viewType: The type of view to register.
+    ///   - method: The view registration method.
     public init(
         reuseIdentifier: String,
         viewType: ViewRegistrationViewType,
@@ -32,7 +42,9 @@ public struct ViewRegistration: Hashable {
         self.method = method
     }
 
-    public func dequeueViewFor(collectionView: UICollectionView, at indexPath: IndexPath) -> UICollectionReusableView {
+    // MARK: Dequeuing Views
+
+    func dequeueViewFor(collectionView: UICollectionView, at indexPath: IndexPath) -> UICollectionReusableView {
         switch self.viewType {
         case .cell:
             return self._dequeueCellFor(collectionView: collectionView, at: indexPath)
@@ -54,7 +66,9 @@ public struct ViewRegistration: Hashable {
         )
     }
 
-    public func registerWith(collectionView: UICollectionView) {
+    // MARK: Registering Views
+
+    func registerWith(collectionView: UICollectionView) {
         switch self.viewType {
         case .cell:
             self._registerCellWith(collectionView: collectionView)
@@ -95,7 +109,19 @@ public struct ViewRegistration: Hashable {
     }
 }
 
+/// Initializes a `ViewRegistration`.
+///
+/// - Parameters:
+///   - reuseIdentifier: The view reuse identifier.
+///   - viewType: The type of view to register.
+///   - method: The view registration method.
 extension ViewRegistration {
+
+    /// Convenience initializer for a class-based cell.
+    ///
+    /// - Parameters:
+    ///   - reuseIdentifier: The cell reuse identifier.
+    ///   - cellClass: The cell class.
     public init(reuseIdentifier: String, cellClass: AnyClass) {
         self.init(
             reuseIdentifier: reuseIdentifier,
@@ -104,6 +130,11 @@ extension ViewRegistration {
         )
     }
 
+    /// Convenience initializer for a nib-based cell in the main bundle.
+    ///
+    /// - Parameters:
+    ///   - reuseIdentifier:  The cell reuse identifier.
+    ///   - cellNibName: The nib name for the cell.
     public init(reuseIdentifier: String, cellNibName: String) {
         self.init(
             reuseIdentifier: reuseIdentifier,
@@ -112,6 +143,12 @@ extension ViewRegistration {
         )
     }
 
+    /// Convenience initializer for a class-based supplementary view.
+    ///
+    /// - Parameters:
+    ///   - reuseIdentifier: The view reuse identifier.
+    ///   - supplementaryViewClass: The view class.
+    ///   - kind: The supplementary view kind.
     public init(reuseIdentifier: String, supplementaryViewClass: AnyClass, kind: String) {
         self.init(
             reuseIdentifier: reuseIdentifier,
@@ -120,6 +157,12 @@ extension ViewRegistration {
         )
     }
 
+    /// Convenience initializer for a nib-based supplementary view in the main bundle.
+    ///
+    /// - Parameters:
+    ///   - reuseIdentifier: The view reuse identifier.
+    ///   - supplementaryViewNibName: The nib name for the view.
+    ///   - kind: The supplementary view kind.
     public init(reuseIdentifier: String, supplementaryViewNibName: String, kind: String) {
         self.init(
             reuseIdentifier: reuseIdentifier,
