@@ -23,7 +23,7 @@ struct FakeCollectionCellViewModel: CellViewModel {
 
     // MARK: CellViewModel
 
-    var id: UniqueIdentifier { self.text }
+    nonisolated var id: UniqueIdentifier { self.text }
 
     func configure(cell: FakeCollectionCell) {
         self.didConfigureExpectation?.fulfill()
@@ -35,15 +35,17 @@ struct FakeCollectionCellViewModel: CellViewModel {
 }
 
 extension XCTestCase {
+    @MainActor
     func makeCollectionViewModel(numSections: Int = 3,
                                  numCells: Int = 5,
                                  includeExpectations: Bool = false) -> CollectionViewModel {
         let sections = (0..<numSections).map { _ in
             self.makeCollectionSectionViewModel(numCells: numCells, includeExpectations: includeExpectations)
         }
-        return CollectionViewModel(sections: sections)
+        return CollectionViewModel(id: "test_model", sections: sections)
     }
 
+    @MainActor
     func makeCollectionSectionViewModel(name: String = .random,
                                         numCells: Int = 5,
                                         includeExpectations: Bool = false) -> SectionViewModel {
@@ -53,6 +55,7 @@ extension XCTestCase {
         return SectionViewModel(id: "section_\(name)", cells: cellModels)
     }
 
+    @MainActor
     func makeCollectionCellViewModel(text: String = .random,
                                      includeExpectations: Bool = false) -> FakeCollectionCellViewModel {
         FakeCollectionCellViewModel(
