@@ -16,8 +16,9 @@ import XCTest
 
 final class TestCellViewModel: XCTestCase {
 
+    @MainActor
     func test_CellViewModel_protocol_default_values() {
-        let viewModel = TestCellModel()
+        let viewModel = FakeCellViewModel()
         XCTAssertTrue(viewModel.shouldHighlight)
     }
 }
