@@ -16,8 +16,9 @@ import XCTest
 
 final class TestDiffableViewModel: XCTestCase {
 
+    @MainActor
     func test_DiffableViewModel_protocol_default_values() {
-        let viewModel = TestCellModel()
-        XCTAssertEqual(viewModel.id, "TestCellModel")
+        let viewModel = FakeCellViewModel()
+        XCTAssertEqual(viewModel.id, "FakeCellViewModel")
     }
 }
