@@ -14,16 +14,17 @@
 @testable import ReactiveCollectionsKit
 import XCTest
 
-class UnitTestCase: XCTestCase {
+open class UnitTestCase: XCTestCase {
 
     private static let frame = CGRect(x: 0, y: 0, width: 320, height: 600)
 
-    let collectionView = FakeCollectionView(
+    @MainActor let collectionView = FakeCollectionView(
         frame: frame,
         collectionViewLayout: FakeCollectionLayout()
     )
 
-    override func setUp() {
+    @MainActor
+    override open func setUp() {
         super.setUp()
         self.collectionView.layoutSubviews()
         self.collectionView.reloadData()
